requirements.txt

streamlit
openai
langdetect
pdfplumber
python-docx
openpyxl
pytesseract
pdf2image
